alphabet: handle large ratio of alphabet to seps
When using a custom alphabet, there may be a large ratio of alphabet
characters to separator characters.  When the ratio is larger than 3.5,
the alphabet and separators are adjusted.

However, if we use integer arithmetic, we may not do the adjustment when
the ratio is > 3.5 but < 4.  Use floating point arithmetic to do this
comparison.

Include a test where the ratio of alphabet characters to separator
characters is 3.714.

+++ src/Hashids.net/Hashids.cs
@@ -224,9 +224,9 @@ namespace HashidsNet
 
             seps = ConsistentShuffle(seps, salt);
 
-            if (seps.Length == 0 || (alphabet.Length / seps.Length) > SEP_DIV)
+            if (seps.Length == 0 || ((float)alphabet.Length / seps.Length) > SEP_DIV)
             {
-                var sepsLength = (int)Math.Ceiling(alphabet.Length / SEP_DIV);
+                var sepsLength = (int)Math.Ceiling((float)alphabet.Length / SEP_DIV);
                 if (sepsLength == 1)
                     sepsLength = 2;
